core, api, hls, srt, rtmp, tests, docs: teardown, ordered revisions, HLS layout
The session's hardening work, in the order it was done.

Teardown and lifetime:
  - registry: each stream owns its pool, released on delete, with a drain gate
    for readers that hold a thread, a refcounted push-destination object, and
    routed slots cleared with the stream they point at
  - hls_pull/hls_ingest: exit when their source is removed (the check had been
    lost, so a deleted source's reader thread never stopped)
  - readers and pools keep the worker's log, not the connection's: keeping a
    request log was a use-after-free, reported by ASan
  - owner directory: a contended lock fails closed instead of entering the
    critical section

Control plane:
  - revisions come from one sequence in shared memory, so two workers' writes
    to one stream are comparable and replicas converge on the higher one
  - deletions are broadcast (the call had been dropped entirely, so a delete
    only ever happened on the worker that answered it) and remembered in a
    bounded tombstone ring, which refuses operations a deletion moved past
  - identity is 64-bit FNV-1a everywhere: owner records, routed slots,
    tombstones and sessions

HLS:
  - one output directory per program, with the segmenter creating the chain and
    the push scan following each program's own directory
  - segments are committed by rename, like the playlist, so a watcher never
    sees a name with a partial segment behind it
  - the input profile (MPEG-TS with H.264/H.265/AAC) is reported when a segment
    carries nothing this build can carry, instead of a source that silently
    never goes on air

HTTP:
  - every blocking wait is bounded: 5 s to connect per address, 10 s without
    progress per read or write, mapped to EAGAIN so a stalled TLS peer is not
    retried forever

Tests and docs:
  - new targets stream-delete and graph-conflict; incarnation is a reproducer
    for an ownership divergence it found, deliberately not in TEST_TARGETS
  - unit suite for the tombstone ring, a commit case in test_hls, a pinned
    FNV-1a 64 vector, and the stale RTMP source-text check removed
  - an ASan build knob (NGINX_CC_OPT/NGINX_LD_OPT) with the recipe documented
  - architecture, api, configuration, operations and development updated

+++ src/core/ngx_media_owner.h
@@ -17,25 +17,36 @@
  * by the bounded internal routing escape hatch (goal doc 23).
  */
 
-/* FNV-1a over "application/stream": stable across processes and builds */
-uint32_t ngx_media_owner_hash(const ngx_str_t *application,
+/*
+ * FNV-1a over "application/stream": stable across processes and builds.
+ *
+ * 64 bits, not 32.  The value is the identity of a stream everywhere the
+ * workers address one - the owner directory's records, the routed slots that
+ * carry a publisher's media to its owner, the tombstones that remember a
+ * deletion - so two distinct streams that hashed alike would be one stream to
+ * the deployment: a frame routed to the wrong program, an owner record shared
+ * by two programs, a delete that removed both.  32 bits collide at a few tens
+ * of thousands of streams by the birthday bound; 64 does not collide in
+ * practice.
+ */
+uint64_t ngx_media_owner_hash(const ngx_str_t *application,
     const ngx_str_t *stream);
 
 /* the worker slot that owns this stream, given the worker count */
-ngx_uint_t ngx_media_owner_slot(uint32_t hash, ngx_uint_t workers);
+ngx_uint_t ngx_media_owner_slot(uint64_t hash, ngx_uint_t workers);
 
 /* the configured worker count (1 when there is no configuration) */
 ngx_uint_t ngx_media_owner_worker_count(ngx_cycle_t *cycle);
 
 /* the slot that owns a stream under the configured worker count */
-ngx_uint_t ngx_media_owner_for(ngx_cycle_t *cycle, uint32_t hash);
+ngx_uint_t ngx_media_owner_for(ngx_cycle_t *cycle, uint64_t hash);
 
 /* the state a worker publishes about a stream it owns */
 #define NGX_MEDIA_OWNER_STATE_FREE      0
 #define NGX_MEDIA_OWNER_STATE_OWNED     1
 #define NGX_MEDIA_OWNER_STATE_DELETED   2
 typedef struct {
-    uint32_t    hash;           /* stream hash this record describes */
+    uint64_t    hash;           /* stream identity this record describes */
     uint32_t    slot;           /* owner worker slot */
     int32_t     pid;            /* owner process id */
     uint64_t    generation;     /* program generation the owner is serving */
